fix crash in SM3Map, minor

+++ cont/base/springcontent/shaders/GLSL/SMFFragProg.glsl
@@ -76,7 +76,7 @@ varying vec2 normalTexCoords;
 varying vec2 infoTexCoords;
 
 #if (HAVE_INFOTEX == 1)
-uniform int infoTexIntensityMul;
+uniform float infoTexIntensityMul;
 #endif
 
 
@@ -206,7 +206,7 @@ void main() {
 		// increase contrast and brightness for the overlays
 		// TODO: make the multiplier configurable by users?
 		diffuseCol.rgb += (texture2D(infoTex, infoTexCoords).rgb * infoTexIntensityMul);
-		diffuseCol.rgb -= (vec3(0.5, 0.5, 0.5) * float(infoTexIntensityMul == 1));
+		diffuseCol.rgb -= (vec3(0.5, 0.5, 0.5) * float(infoTexIntensityMul == 1.0));
 	#endif
